+++ src/background_updates.rs
@@ -1,0 +1,5 @@
+//! Endpoints in the `/_synapse/admin/v<x>/background_updates/` scope to manage background updates.
+
+pub mod enabled;
+pub mod run;
+pub mod status;