+++ src/background_updates/status.rs
@@ -1,0 +1,3 @@
+//! Different versions of the endpoint to get the current status of the background updates.
+
+pub mod v1;